устранение ошибок и доработка предыдущего задания
