+++ src/presenter/presenter.js
@@ -8,15 +8,13 @@ import { POINTS_COUNT } from '../const';
 
 export default class Presentor {
   init = (containers) => {
-    const {controlsContainer, eventsContainer} = containers;
+    const { controlsContainer, eventsContainer } = containers;
     render(new FiltersForm(), controlsContainer);
     render(new SortForm(), eventsContainer);
     render(new EditForm(), eventsContainer);
     render(new AddPointForm(), eventsContainer);
-    let count = 0;
-    while (count < POINTS_COUNT) {
-      render(new TripPoint(), eventsContainer);
-      count++;
-    }
+    Array.from({ length: POINTS_COUNT }, () =>
+      render(new TripPoint(), eventsContainer)
+    );
   };
 }
