+++ src/render.js
@@ -1,18 +1,18 @@
-export const RenderPosition = {
+const RenderPosition = {
   BEFOREBEGIN: 'beforebegin',
   AFTERBEGIN: 'afterbegin',
   BEFOREEND: 'beforeend',
   AFTEREND: 'afterend',
 };
 
-export const createElement = (template) => {
+const createElement = (template) => {
   const newElement = document.createElement('div');
   newElement.innerHTML = template;
 
   return newElement.firstElementChild;
 };
 
-export const render = (
+const render = (
   component,
   container,
   place = RenderPosition.BEFOREEND
@@ -34,3 +34,5 @@ export const render = (
       break;
   }
 };
+
+export {RenderPosition, createElement, render};
