+++ src/view/sorts-view.js
@@ -1,8 +1,8 @@
-import { createSortFormTemplate } from './templates.js';
+import { createSortsFormTemplate } from '../templates/sorts-form.js';
 import { createElement } from '../render.js';
 
 export default class SortForm {
-  getTemplate = () => createSortFormTemplate();
+  getTemplate = () => createSortsFormTemplate();
 
   getElement = () => {
     if (!this.element) {
